requirements.txt
absl-py==1.4.0
accelerate==0.20.3
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
coloredlogs==15.0.1
comm==0.1.3
cycler==0.11.0
Cython==3.0.6
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
diffusers==0.18.2
efficientnet-pytorch==0.7.1
einops==0.6.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
fire==0.5.0
flatbuffers==23.5.26
focal-loss-torch==0.1.2
fonttools==4.40.0
fqdn==1.5.1
fsspec==2023.6.0
fvcore==0.1.5.post20221221
google-auth==2.20.0
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
grpcio==1.54.2
h11==0.14.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
invisible-watermark==0.2.0
iopath==0.1.10
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.23.0
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1rc1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numpy==1.24.3
nuscenes-devkit==1.1.10
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objprint==0.2.3
onnxruntime==1.15.1
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.6.0
pooch==1.7.0
portalocker==2.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyblaze==2.3.2
pycocotools==2.0.6
pycparser==2.21
Pygments==2.15.1
PyMatting==1.1.8
pyparsing==3.1.0
pyquaternion==0.9.9
pyre-extensions==0.0.23
pyro-api==0.1.2
pyro-ppl==1.8.5
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.6.3
rembg==2.0.49
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.10.0
Send2Trash==1.8.2
Shapely==1.8.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tbb==2021.10.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.6.7
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==1.2.0
torchvision==0.14.1+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
trio==0.22.0
trio-websocket==0.10.3
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.16
viztracer==0.16.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wsproto==1.2.0
xformers==0.0.16
yacs==0.1.8
zipp==3.15.0
